Fix crash on presenting ActionSheet for iPad
Terminating app due to uncaught exception 'NSGenericException', reason:
'Your application has presented a UIAlertController of style
UIAlertControllerStyleActionSheet.
The modalPresentationStyle of a UIAlertController with this style is
UIModalPresentationPopover.
You must provide location information for this popover through the alert
controller's popoverPresentationController.
You must provide either a sourceView and sourceRect or a barButtonItem.
If this information is not known when you present the alert controller,
you may provide it in the UIPopoverPresentationControllerDelegate method
-prepareForPopoverPresentation.'